split core.py to make it managable
